채팅 타입 수정
