+++ src/types/chat/index.ts
@@ -1,1 +1,2 @@
 export * from "./chat";
+export * from "./room";
